+++ lib/src/consumer/android.dart
@@ -350,7 +350,7 @@ def localProperties = new Properties()
 |}
 |
 |dependencies {
-|    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.8.20"
+|    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.9.10"
 |}
 |
 |kotlin {